A script with methods dealing with bitarrays
Bitarrays are used to represent the abstractions of states of a chemical system.
So far this script has two methods.
number_of_ones(bitarray) returns the number of ones in a bitarray, i.e. the number of species present in the abstraction represented by the bitstring
lexicographical_position(bitarray) returns the lexicographical position of a bitstring of length n with k ones among all other bitsrings with the same length and amount of ones. This can be used to uniquely assigning positions to abstractions on a horizontal axis which is needed for a hasse-diagramm like representation of the trajectory on a 2-D grid.